Adding ability to upload media
Using the multer module to allow users to upload media to the server. When the media is uploaded, deciding which directory to store it in based on it's mimetype i.e. images get stored in the 'file_uploads/images' directory. Using the custom module I created to check that all of the folders required within the uploads directory exist. If they don't, then they will be created. Envoking the function contained within this module by adding () to the end, as this funciton expects no parametres. Only calling this function when the server starts up, as there should be no reason that this directories would end up being deleted after this point. If I were to check/create these directories each time a file were uploaded, it would significantly increase the time requrired to store the files.