Making the checkDirectories module expandable
The check directories module now accepts an array of directory path names, so that it can be used to check any number of directories. Creating the array for the file uploads directories out in the main app, so that it can be passed into the module, instead of existing only within the checkDirectories module.