Creating a new route to deal with authentication
Imitating a login activity using a predefined username and password (laura and password). This way all request to the /admin section of the site have to pass through the authorisation middleware first, to confirm that they are authorised to access it. Allowing this middle ware to receive requests from /login and /admin, so that when a user tries to login, the authentication portion of that process can also be passed to this middleware.