Adding tabs to homepage, and AJAX checking of username
Using jQuery UI to create a tabbed section on the homepage, so users can create a new portfolio or log in to a current one. Adding an AJAX post request to check if the username the user has requested (in the create new portfolio form) to check if that username is available. Reflecting this response by displaying a glyphicon and status beside the username. Also defaulting the url of the portfolio to be the user's username, but allowing them to change it.