Successfully authorising users using Google OAuth
Created my own custom module to share the google oauth flow between modules. Added in passport sessions to carry the userid as far as the authentication stage, to be set as the username. Currently, google users have a default url of 'googleuser-googleid' where google is relative to their own account. Their username defaults to just their googleid. NOTE - the sub loops in the portfolio.jade view no longer seem to be working, but no change has been made to it since I originally developed that view. Will work on.