Moving the media_uploads directory to the public directory
When I am running this app on Azure (http://portfoliobuilder.azurewebsites.net) the users images cannot be located. I think this may be because I was trying to set the node_modules folder as a static path, when there is already a static path set for the app (/public). Testing this to see if it will work.