Changing the path for the main upload directory
As the media_uploads folders are not currently accessible on the Azure site (as this runs from the bin folder and so the path I was providing to them was not working remotely). Using an enviromnet variable to decide the path, so that it will still work when I am developing locally.