Changing the layout for Login/Google Users
Trying to display the google login as an alternative to creating an account, in a clearer way. Accessing the user's google profile picture, or using the default one I have created, and displaying it on the admin page. General realignment of views, and adding of additional buttons which I will add functionality to later i.e. change profile picture.