Admin can reorder media items
Each time the media items get reordered in the admin panel, and AJAX POST request is sent to the server which contains each media item's id and current index position, so it can be stored in the database. This will then be used to sort the media items next time the admin or portfolio page is loaded.