Fixing Sortable/Editable issue
Up until now, in order to make the title of a media item editable which still being part of the 'sortable' figure element (which can be dragged around to reorder the elements) I was having to call .focus() on the figcaption containing the title, as the sortable() method seemed to be overriding this. Instead, I have now set the cancel option of sortable() to 'figcaption' i.e. if the user clicks while over a figcaption then the sortable() functionality will be cancelled, and hence the figcaption can be edited as normal. The reason I wanted to fix this was because even though the last method (using .focus()) was working in theory, the cursor could not be moved and would default to the start of the title. Also detecting if a user presses the 'enter' key while editing a figcaption, preventing it's default functionality and using it to call .blur() i.e. remove focus on that figcaption.