Linking documents with ObjectId's
Up until now, all media items contained the username of the user which uploaded them in an 'owner' property. Now using ObjectIds to link these, as I feel it is more appropriate. Added a new Portfolio mongoose schema, which will store the pages and ids of media objects which a user has. A new portfolio is created for every user once they create an account. Not yet implemented the linking of media items to portfolios.