Linking Documents and Populating as needed
Using manual referencing to link the user and media items to the portfolio document. Then populating these references when loading a user's portfolio. This means instead of having to query the database for the user, then query the database for their portfolio and then query each of their media items, i can now just query the relevant portfolio and return their owner and all media items in the results.