Rewriting app using Portfolio population
Re-writing the app to work with the portfolio as the primary method of sourcing media items, using population to return the relevant owner and media items in the query result. This method seems much faster, as queries to the database have been greatly reduced by referencing items instead of individually querying them. Also adding a .pre('remove') and a post('save') hook (middleware) to the media items, so that every time a media item is deleted, its object id will also be deleted from the portfolio it exists in, and everytime a new media item is saved, its id will be added to the owners portfolio and stored there.