Persisting tabs and accordions between page loads
So that the tab or accordion option that was open prior to the page being reloaded (due to a file upload, or user refreshing the page) will be reopened after the page reloads. Also opening the appropriate pages tab in the admin panel, based on which file the user is uploading. i.e. if a user uploads a new contact page image, the contact tab will be open once the file has been uploaded.