Fixing sorting of media items
Changing the sort function of media items to be a property on the exports of the databaseModels, as opposed to a method on the Portfolio model. The purpose of this is that now that the user's portfolio is being stored on the session (when they are in the admin panel) I now need to be able to call the sortMediaItems() method on the session media items, as well as the ones coming straight from the portfolio (e.g. on the portfolio page).