Validating requests
Working to improve the authentication process of the router, as I realised that only GET request were be authenticated in the authentication route. Instead, now cathing all request types and checking if a user is logged in before they are allowed to reach the admin section. Adding a filter to multer, so that files can only be uploaded by logged in users, and only certain file types will be accepted. (Testing this theory by deleting the users session cookie in the browser and trying to upload a file. Up until now, the file wouldn't be added to their database due to the authentication step, but it would be stored to the server, as the Multer middleware exists outside of the authentication route). Now checking if the user is logged in before the file is even given a path to be stored to, and rejecting the file if no user is detected or the file type is not accepted by the app.