No longer storing portfolio on session
After much thought and testing, I have decided to revert back to my original method of updating the database directly each time a change is made to a users portfolio. There are a number of reasons for this - 1) If I were to login to my portfolio on two different devices at the same time, the two 'portfolio's I would be working off quickly become out of sync (as they only reflect the state of the portfolio when their session was created, and hence changes made on one device won't be reflected on the other until they log out, and their session becomes obsolete). 2 - I am now storing sessions to the databases for production, which means that every time I make a change to the session object it is making a call to the database anyway. 3 - Because the sessions are now stored in the database, having the portfolio on them means that the owner, portfolio and media items of each user could be stored multiple times in the same database (i.e. I logged in on two devices, and could see that the database now contained the original portfolio, owner and media items, as well as two additional sessions which contained a duplicate of this data). 4 - By working directly with the database, the amount of code required for each request is greatly reduced. 5 - I have designed the admin panel so that the majority of requests are handled through AJAX, so there is no noticable time delay for the user (except for when they upload a media file). From what I have been able to trace out in testing, the greatest delay is the time it takes for the file to be fully uploaded to the server, and once this is successful, the time it takes to store it in the database is minimal. NOTE - prior to this commit, the admin panel was fully functional on the session object for getting data from the portfolio.