Restricting file upload to 1 file at a time
Due to the delays when the app in running on Azure, I feel this is a safer option as there is less chance of the request timing out with one file. The code on the server still provides for multiple file uploads at once, I am just not allowing them at present.